backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
itsdangerous==2.2.0
Werkzeug==3.1.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
